[CMake] Prefer SDL2 in "external" folder and don't look at systems paths
On Linux without "NO_CMAKE_SYSTEM_PATH", CMake would choose the SDL2 library from "/usr/lib" or whatever the system path is.

If the computer is running Linux on "amd64"/"x86_64" arch, it would pick a 64 bits version of the SDL2 library. This would lead to a 64 bits library being used in a 32 bits application.

"NO_CMAKE_SYSTEM_PATH" prevent this to happen and will look at HL:E's folders and finally Half-Life's for the proper SDL2 library to use.